refactor: merge tool service into stage tool
