+++ src/services/index.js
@@ -6,7 +6,6 @@ import { useQueryService } from './query';
 import { useSelectService } from './select';
 import { useStageService } from './stage';
 import { useStageToolService } from './stageTool';
-import { useToolService } from './tool';
 import { useViewportService } from './viewport';
 
 export {
@@ -18,7 +17,6 @@ export {
     useSelectService,
     useStageService,
     useStageToolService,
-    useToolService,
     useViewportService
 };
 
@@ -31,6 +29,5 @@ export const useService = () => ({
     select: useSelectService(),
     stage: useStageService(),
     stageTool: useStageToolService(),
-    tool: useToolService(),
     viewport: useViewportService()
 });
